To enhance the reuseability of code, move function 'tcmu_cancel_thread' from 'main.c' to 'api.c'
